move command tests..
